requirements.txt
python-telegram-bot==20.3
flask
openai
ffmpeg-python
requests
